refactored page builder to allow factories
