+++ lib/page_object/page_builder/page_builder.rb
@@ -1,0 +1,54 @@
+require_relative '../page'
+
+module PageObject
+  module PageBuilder
+    class PageBuilder
+      class << self
+        attr_reader :stack, :factories
+
+        def define(&block)
+          @factories ||= {}
+          #reset the stack
+          @stack = []
+          #place our hash
+          @stack.push Hash.new
+          #run the block given
+          instance_eval &block if block_given?
+          #take the top element and pass as the page's hash
+          Page.new @stack.pop
+        end#define
+
+        def selector(hash)
+          @stack.last[:selector] ||= []
+          @stack.last[:selector].push Selector.new hash
+        end#selector
+
+        def actions(hash)
+          @stack.last[:actions] ||= {}
+          @stack.last[:actions].merge!(hash) {|_key, _oldval, newval| newval}
+        end#actions
+
+        def next_page(name)
+          @stack.last[:next_page] = name
+        end#next_page
+
+        def check(hash)
+          method_value = hash.shift
+          @stack.last[:before] = proc {|driver| driver.send(method_value[0]).match method_value[1]}
+        end
+
+        def factories
+          @factories ||= {}
+        end#factories
+
+        def method_missing(method, *args, &block)
+          if @factories.has_key? method
+            @factories[method].send method, *args, &block
+          else
+            super
+          end#else
+        end#method_missing
+      end#class << self
+    end#class PageBuilder
+  end#module PageBuilder
+end#PageObject
